+++ CHANGELOG.md
@@ -230,5 +230,5 @@
 
 ## v0.1.7 2025-11-21
 
-- Initial implementation of OpenSpec for Copilot features.
+- Initial implementation of OpenSpec for Agent features.